requirements_full.txt
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
Brotli==1.1.0
build==1.0.3
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.7
coloredlogs==15.0.1
contextlib2==21.6.0
crewai==0.16.3
crewai-tools==0.0.12
cryptography==42.0.5
curl_cffi==0.6.1
dataclasses-json==0.6.4
decorator==5.1.1
Deprecated==1.2.14
deprecation==2.1.0
distro==1.9.0
docstring-parser==0.15
duckduckgo_search==4.5.0
embedchain==0.1.88
exceptiongroup==1.2.0
fastapi==0.110.0
filelock==3.13.1
flatbuffers==23.5.26
frozenlist==1.4.1
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.42
google-api-core==2.17.1
google-auth==2.28.1
google-cloud-aiplatform==1.42.1
google-cloud-bigquery==3.17.2
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.2
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
gptcache==0.1.43
greenlet==3.0.3
grpc-google-iam-v1==0.13.0
grpcio==1.62.0
grpcio-status==1.62.0
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.21.2
humanfriendly==10.0
idna==3.6
importlib-metadata==6.11.0
importlib_resources==6.1.2
iniconfig==2.0.0
instructor==0.5.2
jsonpatch==1.33
jsonpointer==2.4
kubernetes==29.0.0
lancedb==0.5.7
langchain==0.1.9
langchain-community==0.0.24
langchain-core==0.1.27
langchain-openai==0.0.5
langsmith==0.1.10
lxml==5.1.0
Mako==1.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.0
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
mutagen==1.47.0
mypy-extensions==1.0.0
nodeenv==1.8.0
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.17.1
openai==1.12.0
opentelemetry-api==1.23.0
opentelemetry-exporter-otlp-proto-common==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-exporter-otlp-proto-http==1.23.0
opentelemetry-instrumentation==0.44b0
opentelemetry-instrumentation-asgi==0.44b0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-proto==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
opentelemetry-util-http==0.44b0
orjson==3.9.15
overrides==7.7.0
packaging==23.2
pluggy==1.4.0
posthog==3.4.2
proto-plus==1.23.0
protobuf==4.25.3
pulsar-client==3.4.0
py==1.11.0
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodomex==3.20.0
pydantic==2.6.3
pydantic_core==2.16.3
PyGithub==1.59.1
Pygments==2.17.2
PyJWT==2.8.0
pylance==0.9.18
PyNaCl==1.5.0
pypdf==3.17.4
PyPika==0.48.9
pyproject_hooks==1.0.0
pyright==1.1.351
pysbd==0.3.4
pytest==8.0.2
python-dateutil==2.8.2
python-dotenv==1.0.1
pytube==15.0.0
PyYAML==6.0.1
ratelimiter==1.2.0.post0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
retry==0.9.2
rich==13.7.1
rsa==4.9
schema==0.7.5
semver==3.0.2
shapely==2.0.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.27
starlette==0.36.3
sympy==1.12
tenacity==8.2.3
tiktoken==0.5.2
tokenizers==0.15.2
tomli==2.0.1
tqdm==4.66.2
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.27.1
uvloop==0.19.0
watchfiles==0.21.0
websocket-client==1.7.0
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
youtube-transcript-api==0.6.2
yt-dlp==2023.12.30
zipp==3.17.0
